+++ packages/app/app/components/Header.tsx
@@ -6,7 +6,7 @@ const Header = () => {
     <header className="py-8 border-b mb-10">
       <Wrapper>
         <div className="flex items-center justify-between">
-          <h1 className="text-lg md:text-2xl font-bold">
+          <h1 className="text-lg md:text-xl font-bold">
             Solidity Next.js Starter
           </h1>
           <ConnectButton