+++ packages/app/app/components/Providers.tsx
@@ -1,38 +1,26 @@
 "use client";
 
 import { ReactNode } from "react";
-import { WagmiConfig, configureChains, createConfig } from "wagmi";
-import { getDefaultWallets, RainbowKitProvider } from "@rainbow-me/rainbowkit";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { WagmiProvider } from "wagmi";
+import { getDefaultConfig, RainbowKitProvider } from "@rainbow-me/rainbowkit";
 import { hardhat, goerli } from "wagmi/chains";
-import { alchemyProvider } from "wagmi/providers/alchemy";
-import { publicProvider } from "wagmi/providers/public";
 
-const chain = process.env.NEXT_PUBLIC_CHAIN_ID === "5" ? goerli : hardhat;
-const { chains, publicClient, webSocketPublicClient } = configureChains(
-  [chain],
-  [
-    alchemyProvider({
-      apiKey: process.env.NEXT_PUBLIC_ALCHEMY_API_KEY ?? "",
-    }),
-    publicProvider(),
-  ],
-);
-const { connectors } = getDefaultWallets({
-  appName: "Solidity Next.js Starter",
+const queryClient = new QueryClient();
+
+const config = getDefaultConfig({
+  appName: 'Solidity Next.js Starter',
   projectId: process.env.NEXT_PUBLIC_RAINBOWKIT_PROJECT_ID ?? "",
-  chains,
-});
-const config = createConfig({
-  autoConnect: true,
-  connectors,
-  publicClient,
-  webSocketPublicClient,
+  chains: [hardhat, goerli],
+  ssr: true,
 });
 
 const Providers = ({ children }: { children: ReactNode }) => (
-  <WagmiConfig config={config}>
-    <RainbowKitProvider chains={chains}>{children}</RainbowKitProvider>
-  </WagmiConfig>
+  <WagmiProvider config={config}>
+    <QueryClientProvider client={queryClient}>
+      <RainbowKitProvider>{children}</RainbowKitProvider>
+    </QueryClientProvider>
+  </WagmiProvider>
 );
 
 export { Providers };